Inserting #(squash) layer last
Previously, the squash layer was inserted immediately after the root.
The reason for doing this initially was to follow the logic in the
original `docker-squash` library for lack of a better place to start.
However, now that this library reflects a better understand of the
docker image format, it has become apparent that there is no real reason
to insert the layer after the root.  Instead, inserting the layer at the
end has the advantage of placing the layers in a more logical order.  The squash is
the last operation done on the image, so putting the squash layer last
instead of somewhere in the middle makes the image history less
confusing.

Additionally, it is now possible to do multiple concecutive builds,
squashing after each, and producing a working image.  Previously, the
logic would have prevented any previous squash layers from being
squashed into the new squash layer.

Finally, this pull request changes the process of rewriting the
layer children.  Whereas previously the `ADD` and `RUN` layers were
completely removed, now they are kept as-is except that their
`layer.tar` is zeroed out.  This allows for the more accurate
preservation of layer history, which is potentially useful for debugging
purposes.